added hero image
