+++ src/components/shared/container/Container.tsx
@@ -6,7 +6,5 @@ interface ContainerProps {
 }
 
 export default function Container({ children, className }: ContainerProps) {
-  return (
-    <div className={`container tab:max-w-full ${className}`}>{children}</div>
-  );
+  return <div className={`container max-w-full ${className}`}>{children}</div>;
 }
